BE/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
python-multipart
python-dotenv
pandas
openpyxl
email-validator